Move client options to a new file.
